+++ packages/luciq_flutter/ios/Classes/Util/LCQAPM+PrivateAPIs.h
@@ -8,6 +8,7 @@
 
 #import <LuciqSDK/LCQAPM.h>
 #import "LCQTimeIntervalUnits.h"
+#import <LuciqSDK/LCQFrameInfo.h>
 
 @interface LCQAPM (PrivateAPIs)
 
@@ -15,11 +16,21 @@
 /// `endScreenLoadingEnabled` will be only true if  APM, screenLoadingFeature.enabled and autoUITracesUserPreference are true
 @property (class, atomic, assign) BOOL endScreenLoadingEnabled;
 
++ (void)setScreenRenderingEnabled:(BOOL)enabled;
+
 + (void)startUITraceCPWithName:(NSString *)name startTimestampMUS:(LCQMicroSecondsTimeInterval)startTimestampMUS;
 
 + (void)reportScreenLoadingCPWithStartTimestampMUS:(LCQMicroSecondsTimeInterval)startTimestampMUS
                                        durationMUS:(LCQMicroSecondsTimeInterval)durationMUS;
 
 + (void)endScreenLoadingCPWithEndTimestampMUS:(LCQMicroSecondsTimeInterval)endTimestampMUS;
+
++ (BOOL)isScreenRenderingOperational;
+
++ (void)endAutoUITraceCPWithFrames:(nullable NSArray<LCQFrameInfo *> *)frames;
+
++ (void)endCustomUITraceCPWithFrames:(nullable NSArray<LCQFrameInfo *> *)frames;
+
++ (double)screenRenderingThreshold;
 
 @end