+++ packages/luciq_dio_interceptor/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [3.1.0](https://github.com/luciqai/luciq-flutter-sdk/compare/v19.0.0...dev) (December 17, 2025)
+
+- Support LUCIQ SDK v19.x.x
+
 ## [3.0.0](https://github.com/luciqai/luciq-flutter-sdk/compare/v18.0.0...dev) (September 24, 2025)
 
 - SDK rebranded from Instabug to Luciq.